+++ tests/core/test_views.py
@@ -1,9 +1,30 @@
-import pytest
 from unittest import mock
+import json
+
+import pytest
 
 from django.urls import reverse
 
 from core import helpers
+from tests.helpers import create_response
+
+
+@pytest.fixture
+def enrol_data():
+    return {
+        'company_name': 'Example corp',
+        'expertise_industries': json.dumps({'name': 'Science'}),
+        'expertise_countries': json.dumps(['USA']),
+        'first_name': 'foo',
+        'last_name': 'bar',
+    }
+
+
+@pytest.fixture(autouse=True)
+def mock_get_company_profile():
+    stub = mock.patch('sso.helpers.get_company_profile', return_value=None)
+    yield stub.start()
+    stub.stop()
 
 
 @pytest.mark.django_db
@@ -59,3 +80,61 @@ def test_exportplan_view(mock_get_export_plan_rules_regs, mock_user_location_cre
     assert mock_get_export_plan_rules_regs.call_args == mock.call(sso_session_id=user.session_id,)
 
     assert response.context['rules_regulation'] == {'rule1': 'r1'}
+
+
+@pytest.mark.django_db
+@mock.patch.object(helpers, 'create_company_profile')
+@mock.patch.object(helpers, 'create_user_profile')
+def test_api_create_company_success(mock_create_user_profile, mock_create_company_profile, client, user, enrol_data):
+    client.force_login(user)
+
+    mock_create_user_profile.return_value = create_response()
+    mock_create_company_profile.return_value = create_response()
+
+    response = client.post(reverse('core:api-create-company'), enrol_data)
+    assert response.status_code == 200
+    assert mock_create_user_profile.call_count == 1
+    assert mock_create_user_profile.call_args == mock.call(
+        data={'first_name': 'foo', 'last_name': 'bar'},
+        sso_session_id=user.session_id,
+    )
+    assert mock_create_company_profile.call_count == 1
+    assert mock_create_company_profile.call_args == mock.call({
+        'sso_id': user.id,
+        'company_name': enrol_data['company_name'],
+        'expertise_industries': enrol_data['expertise_industries'],
+        'expertise_countries': enrol_data['expertise_countries'],
+        'company_email': user.email,
+        'contact_email_address': user.email,
+    })
+
+
+@pytest.mark.django_db
+@mock.patch.object(helpers, 'create_company_profile')
+@mock.patch.object(helpers, 'create_user_profile')
+def test_api_create_company_validation_error(mock_create_user_profile, mock_create_company_profile, client, user):
+    client.force_login(user)
+
+    response = client.post(reverse('core:api-create-company'), {})
+
+    assert response.status_code == 400
+    assert mock_create_user_profile.call_count == 0
+    assert mock_create_company_profile.call_count == 0
+
+
+@pytest.mark.django_db
+def test_api_create_company_not_logged_in(client, enrol_data):
+    response = client.post(reverse('core:api-create-company'), enrol_data)
+
+    assert response.status_code == 403
+
+
+@pytest.mark.django_db
+def test_api_create_company_already_has_company(mock_get_company_profile, client, user, enrol_data):
+    mock_get_company_profile.return_value = {'foo': 'bar'}
+
+    client.force_login(user)
+
+    response = client.post(reverse('core:api-create-company'), enrol_data)
+
+    assert response.status_code == 403