+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - MVP-111 - Show welcome message to logged in users
 - MVP-125 - Store user's location
 - MVP-102 - select-commodity-code 
+- MVP-146 - Question modal